Rename the Identities component to make it Participants generic
